+++ app/Services/QuestionService.php
@@ -4,8 +4,10 @@ namespace App\Services;
 
 use App\Models\Comment;
 use App\Models\Question;
+use App\Models\QuestionBank;
 use Exception;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionService
 {
@@ -25,6 +27,7 @@ class QuestionService
   public function getNonAssignQuestions()
   {
     $user = auth()->user();
+
     return Question::with(['subject:name,id', 'teacher:id,fullname'])
       ->where('subject_id', $user->subject_id)
       ->whereNull('review_by')
@@ -54,7 +57,12 @@ class QuestionService
 
   public function updateQuestion(Question $question, array $data)
   {
+    if (!Gate::allows('can-update-question', $question)) {
+      abort(403);
+    }
+
     $question->update($data);
+    
     $question->answers()->delete();
     foreach ($data['answers'] as $key => $answer) {
       $answer = [
@@ -68,25 +76,33 @@ class QuestionService
 
   public function deleteQuestion(Question $question)
   {
-    $question->answers()->delete();
-    $question->delete();
+    if (!Gate::allows('can-update-question', $question)) {
+      abort(403);
+    }
+    
+    return $question->delete();
   }
 
   public function getAllCommentForQuestion(Question $question)
   {
-    return Comment::whereQuestionId($question->id)->whereNull('comment_id')->with('commentor:id,fullname')->latest()->get();
+    return Comment::whereQuestionId($question->id)->whereNull('comment_id')
+      ->with('commentor:id,fullname,avatar')->orderBy('created_at', 'asc')->get();
   }
 
-  public function assign(array $questions, $teacher_id)
+  public function assign(string $questions, $teacher_id)
   {
-    DB::transaction(function () use ($questions, $teacher_id) {
-      foreach ($questions as $question) {
-        Question::find($question)->update([
-          'review_by' => $teacher_id,
-          'status' => QUESTION_STATUS_WAITING_REVIEW
-        ]);
-      }
-    });
+    $questions = explode(',', $questions);
+    try {
+      DB::beginTransaction();
+      Question::whereIn('id', $questions)->update([
+        'review_by' => $teacher_id,
+        'status' => QUESTION_STATUS_WAITING_REVIEW
+      ]);
+      DB::commit();
+    } catch (Exception $e) {
+      DB::rollBack();
+      throw $e;
+    }
   }
 
   public function reviewQuestion(Question $question, $status)
@@ -94,5 +110,20 @@ class QuestionService
     $question->update([
       'status' => $status
     ]);
+  }
+
+  public function acceptQuestionToBank(Question $question)
+  {
+    // Check question is not already in bank
+    // $alreadyQuestions = QuestionBank::searchByQuery(['match' => ['content' => $question->content]]);
+    $alreadyQuestions = QuestionBank::whereIn('id', [1, 2, 3])->get();
+    
+    if (count($alreadyQuestions) > 0) {
+      throw new Exception('Question is already in bank');
+    }
+    
+    $question->load('answers');
+    $questionBank = QuestionBank::create($question->toArray());
+    $questionBank->answers()->create($question->answers->toArray());
   }
 }
